+++ src/components/navigation/HeaderBackground.tsx
@@ -4,7 +4,7 @@ import LinearGradient from 'react-native-linear-gradient'
 import Animated, { interpolate, SharedValue, useAnimatedStyle } from 'react-native-reanimated'
 
 import { useSceneScrollContext } from '../../state/SceneScrollState'
-import { BlurBackground } from '../common/BlurBackground'
+import { BlurBackgroundNoRoundedCorners } from '../common/BlurBackground'
 import { styled } from '../hoc/styled'
 import { useTheme } from '../services/ThemeContext'
 import { DividerLine } from '../themed/DividerLine'
@@ -17,7 +17,7 @@ export const HeaderBackground = (props: any) => {
 
   return (
     <HeaderBackgroundContainerView scrollY={scrollY}>
-      <BlurBackground />
+      <BlurBackgroundNoRoundedCorners />
       <HeaderLinearGradient colors={theme.headerBackground} start={theme.headerBackgroundStart} end={theme.headerBackgroundEnd} />
       <DividerLine colors={theme.headerOutlineColors} />
     </HeaderBackgroundContainerView>