+++ src/constants/WalletAndCurrencyConstants.ts
@@ -160,6 +160,7 @@ interface SpecialCurrencyInfo {
   maxSpendTargets?: number
   walletConnectV2ChainId?: WalletConnectChainId
   chainIcon?: boolean
+  unstoppableDomainsTicker?: string // https://support.unstoppabledomains.com/support/solutions/articles/48001185621
 }
 
 /*
@@ -197,7 +198,8 @@ export const SPECIAL_CURRENCY_INFO: {
     displayBuyCrypto: true,
     displayIoniaRewards: true,
     isImportKeySupported: true,
-    isStakingSupported: true
+    isStakingSupported: true,
+    unstoppableDomainsTicker: 'BTC'
   },
   bitcointestnet: {
     hasSegwit: true,
@@ -214,21 +216,24 @@ export const SPECIAL_CURRENCY_INFO: {
     displayBuyCrypto: true,
     displayIoniaRewards: true,
     isImportKeySupported: true,
-    isStakingSupported: true
+    isStakingSupported: true,
+    unstoppableDomainsTicker: 'BCH'
   },
   bitcoinsv: {
     maxSpendTargets: UTXO_MAX_SPEND_TARGETS,
     initWalletName: lstrings.string_first_bitcoin_sv_wallet_name,
     chainCode: 'BSV',
     keysOnlyMode: true,
-    isImportKeySupported: true
+    isImportKeySupported: true,
+    unstoppableDomainsTicker: 'BSV'
   },
   digibyte: {
     maxSpendTargets: UTXO_MAX_SPEND_TARGETS,
     initWalletName: lstrings.string_first_digibyte_wallet_name,
     chainCode: 'DGB',
     displayBuyCrypto: true,
-    isImportKeySupported: true
+    isImportKeySupported: true,
+    unstoppableDomainsTicker: 'DGB'
   },
   litecoin: {
     hasSegwit: true,
@@ -238,7 +243,8 @@ export const SPECIAL_CURRENCY_INFO: {
     displayBuyCrypto: true,
     displayIoniaRewards: true,
     isImportKeySupported: true,
-    isStakingSupported: true
+    isStakingSupported: true,
+    unstoppableDomainsTicker: 'LTC'
   },
   rsk: {
     initWalletName: lstrings.string_first_rsk_wallet_name,
@@ -249,7 +255,8 @@ export const SPECIAL_CURRENCY_INFO: {
     walletConnectV2ChainId: {
       namespace: 'eip155',
       reference: '30'
-    }
+    },
+    unstoppableDomainsTicker: 'RSK'
   },
   stellar: {
     initWalletName: lstrings.string_first_stellar_wallet_name,
@@ -261,7 +268,8 @@ export const SPECIAL_CURRENCY_INFO: {
       alertMessage: lstrings.request_xlm_minimum_notification_alert_body
     },
     displayBuyCrypto: false,
-    isImportKeySupported: true
+    isImportKeySupported: true,
+    unstoppableDomainsTicker: 'XLM'
   },
   ripple: {
     initWalletName: lstrings.string_first_ripple_wallet_name,
@@ -275,13 +283,15 @@ export const SPECIAL_CURRENCY_INFO: {
     },
     displayBuyCrypto: false,
     tokenActivationAdditionalReserveText: lstrings.activate_wallet_token_scene_body_xrp_extra,
-    isImportKeySupported: true
+    isImportKeySupported: true,
+    unstoppableDomainsTicker: 'XRP'
   },
   monero: {
     initWalletName: lstrings.string_first_monero_wallet_name,
     chainCode: 'XMR',
     dummyPublicAddress: '46qxvuS78CNBoiiKmDjvjd5pMAZrTBbDNNHDoP52jKj9j5mk6m4R5nU6BDrWQURiWV9a2n5Sy8Qo4aJskKa92FX1GpZFiYA',
-    isImportKeySupported: false
+    isImportKeySupported: false,
+    unstoppableDomainsTicker: 'XMR'
   },
   cardano: {
     initWalletName: lstrings.string_first_cardano_wallet_name,
@@ -290,7 +300,8 @@ export const SPECIAL_CURRENCY_INFO: {
     noMaxSpend: true,
     dummyPublicAddress: 'addr1qyh498v7479sljadw8mdlmshnlt3n30ewzpqnmvrsz2v8rpqt56tgy6jhzgcc7v8mlh7lhw9a9j2hdlmek4arx2238us9e5fq0',
     isImportKeySupported: true,
-    isStakingSupported: true
+    isStakingSupported: true,
+    unstoppableDomainsTicker: 'ADA'
   },
   cardanotestnet: {
     initWalletName: lstrings.string_first_cardano_preprod_wallet_name,
@@ -308,7 +319,8 @@ export const SPECIAL_CURRENCY_INFO: {
     isImportKeySupported: true,
     keysOnlyMode: true,
     needsAccountNameSetup: true,
-    noChangeMiningFee: true
+    noChangeMiningFee: true,
+    unstoppableDomainsTicker: 'EOS'
   },
   telos: {
     initWalletName: lstrings.string_first_telos_wallet_name,
@@ -318,7 +330,8 @@ export const SPECIAL_CURRENCY_INFO: {
     dummyPublicAddress: 'edgecreator2',
     needsAccountNameSetup: true,
     noChangeMiningFee: true,
-    isImportKeySupported: true
+    isImportKeySupported: true,
+    unstoppableDomainsTicker: 'TLOS'
   },
   wax: {
     initWalletName: lstrings.string_first_wax_wallet_name,
@@ -328,7 +341,8 @@ export const SPECIAL_CURRENCY_INFO: {
     needsAccountNameSetup: false,
     noChangeMiningFee: true,
     isImportKeySupported: true,
-    keysOnlyMode: true
+    keysOnlyMode: true,
+    unstoppableDomainsTicker: 'WAXP'
   },
   ethereum: {
     initWalletName: lstrings.string_first_ethereum_wallet_name,
@@ -341,7 +355,8 @@ export const SPECIAL_CURRENCY_INFO: {
     walletConnectV2ChainId: {
       namespace: 'eip155',
       reference: '1'
-    }
+    },
+    unstoppableDomainsTicker: 'ETH'
   },
   arbitrum: {
     initWalletName: lstrings.string_first_arbitrum_wallet_name,
@@ -374,7 +389,8 @@ export const SPECIAL_CURRENCY_INFO: {
     displayBuyCrypto: false,
     isImportKeySupported: true,
     isStakingSupported: false,
-    noMaxSpend: true
+    noMaxSpend: true,
+    unstoppableDomainsTicker: 'FIL'
   },
   filecoinfevm: {
     initWalletName: lstrings.string_first_filecoin_fevm_wallet_name,
@@ -399,7 +415,8 @@ export const SPECIAL_CURRENCY_INFO: {
     allowZeroTx: true,
     noChangeMiningFee: true,
     isImportKeySupported: true,
-    isStakingSupported: true
+    isStakingSupported: true,
+    unstoppableDomainsTicker: 'TRX'
   },
   ethereumclassic: {
     initWalletName: lstrings.string_first_ethereum_classic_wallet_name,
@@ -409,7 +426,8 @@ export const SPECIAL_CURRENCY_INFO: {
     walletConnectV2ChainId: {
       namespace: 'eip155',
       reference: '61'
-    }
+    },
+    unstoppableDomainsTicker: 'ETC'
   },
   ethereumpow: {
     initWalletName: lstrings.string_first_ethereum_pow_wallet_name,
@@ -466,7 +484,8 @@ export const SPECIAL_CURRENCY_INFO: {
     // will share / copy public address instead of URI on Request scene
     isUriEncodedStructure: true,
     dummyPublicAddress: 'tz1cVgSd4oY25pDkH7vdvVp5DfPkZwT2hXwX',
-    isImportKeySupported: true
+    isImportKeySupported: true,
+    unstoppableDomainsTicker: 'XTZ'
   },
   axelar: {
     initWalletName: lstrings.string_first_axelar_wallet_name,
@@ -499,7 +518,8 @@ export const SPECIAL_CURRENCY_INFO: {
     walletConnectV2ChainId: {
       namespace: 'cosmos',
       reference: 'cosmoshub-4'
-    }
+    },
+    unstoppableDomainsTicker: 'ATOM'
   },
   osmosis: {
     initWalletName: lstrings.string_first_osmosis_wallet_name,
@@ -516,14 +536,16 @@ export const SPECIAL_CURRENCY_INFO: {
     chainCode: 'TON',
     dummyPublicAddress: 'UQAc_4sYewa5e5eN1D3nrt9wDy2akCCQ3VyNlhcxF4VozlO5',
     isImportKeySupported: false,
-    noChangeMiningFee: true
+    noChangeMiningFee: true,
+    unstoppableDomainsTicker: 'TON'
   },
   thorchainrune: {
     initWalletName: lstrings.string_first_thorchainrune_wallet_name,
     chainCode: 'RUNE',
     noChangeMiningFee: true,
     dummyPublicAddress: 'thor1mj5j3eke6m9tcvmn8lwwxdrputyvax45lqawch',
-    isImportKeySupported: true
+    isImportKeySupported: true,
+    unstoppableDomainsTicker: 'RUNE'
   },
   binance: {
     initWalletName: lstrings.string_first_bnb_wallet_name,
@@ -549,7 +571,8 @@ export const SPECIAL_CURRENCY_INFO: {
     chainCode: 'SOL',
     isImportKeySupported: true,
     dummyPublicAddress: 'DEd1rkRyr5bRkJHgaAKMSYjYC1KMz3Hc5bSs4Jiwt29x',
-    noChangeMiningFee: true
+    noChangeMiningFee: true,
+    unstoppableDomainsTicker: 'SOL'
   },
   celo: {
     initWalletName: lstrings.string_first_celo_wallet_name,
@@ -560,7 +583,8 @@ export const SPECIAL_CURRENCY_INFO: {
     walletConnectV2ChainId: {
       namespace: 'eip155',
       reference: '42220'
-    }
+    },
+    unstoppableDomainsTicker: 'CELO'
   },
   fio: {
     allowZeroTx: true,
@@ -577,13 +601,15 @@ export const SPECIAL_CURRENCY_INFO: {
     initWalletName: lstrings.string_first_dash_wallet_name,
     chainCode: 'DASH',
     displayIoniaRewards: true,
-    isImportKeySupported: true
+    isImportKeySupported: true,
+    unstoppableDomainsTicker: 'DASH'
   },
   ravencoin: {
     maxSpendTargets: UTXO_MAX_SPEND_TARGETS,
     initWalletName: lstrings.string_first_ravencoin_wallet_name,
     chainCode: 'RVN',
-    isImportKeySupported: true
+    isImportKeySupported: true,
+    unstoppableDomainsTicker: 'RVN'
   },
   dogecoin: {
     maxSpendTargets: UTXO_MAX_SPEND_TARGETS,
@@ -591,20 +617,23 @@ export const SPECIAL_CURRENCY_INFO: {
     chainCode: 'DOGE',
     displayIoniaRewards: true,
     isImportKeySupported: true,
-    isStakingSupported: true
+    isStakingSupported: true,
+    unstoppableDomainsTicker: 'DOGE'
   },
   zcoin: {
     maxSpendTargets: UTXO_MAX_SPEND_TARGETS,
     initWalletName: lstrings.string_first_zcoin_wallet_name,
     chainCode: 'FIRO',
-    isImportKeySupported: true
+    isImportKeySupported: true,
+    unstoppableDomainsTicker: 'FIRO'
   },
   smartcash: {
     maxSpendTargets: UTXO_MAX_SPEND_TARGETS,
     initWalletName: lstrings.string_first_smartcash_wallet_name,
     chainCode: 'SMART',
     isImportKeySupported: true,
-    keysOnlyMode: true
+    keysOnlyMode: true,
+    unstoppableDomainsTicker: 'SMART'
   },
   vertcoin: {
     maxSpendTargets: UTXO_MAX_SPEND_TARGETS,
@@ -617,7 +646,8 @@ export const SPECIAL_CURRENCY_INFO: {
     initWalletName: lstrings.string_first_bitcoin_gold_wallet_name,
     chainCode: 'BTG',
     isImportKeySupported: true,
-    isSplittingDisabled: true
+    isSplittingDisabled: true,
+    unstoppableDomainsTicker: 'BTG'
   },
   feathercoin: {
     maxSpendTargets: UTXO_MAX_SPEND_TARGETS,
@@ -629,13 +659,15 @@ export const SPECIAL_CURRENCY_INFO: {
     maxSpendTargets: UTXO_MAX_SPEND_TARGETS,
     initWalletName: lstrings.string_first_groestlcoin_wallet_name,
     chainCode: 'GRS',
-    isImportKeySupported: true
+    isImportKeySupported: true,
+    unstoppableDomainsTicker: 'GRS'
   },
   qtum: {
     maxSpendTargets: UTXO_MAX_SPEND_TARGETS,
     initWalletName: lstrings.string_first_qtum_wallet_name,
     chainCode: 'QTUM',
-    isImportKeySupported: true
+    isImportKeySupported: true,
+    unstoppableDomainsTicker: 'QTUM'
   },
   eboost: {
     maxSpendTargets: UTXO_MAX_SPEND_TARGETS,
@@ -659,7 +691,8 @@ export const SPECIAL_CURRENCY_INFO: {
     walletConnectV2ChainId: {
       namespace: 'eip155',
       reference: '250'
-    }
+    },
+    unstoppableDomainsTicker: 'FTM'
   },
   hedera: {
     initWalletName: lstrings.string_first_hedera_wallet_name,
@@ -667,7 +700,8 @@ export const SPECIAL_CURRENCY_INFO: {
     dummyPublicAddress: '0.0.14625',
     isImportKeySupported: true,
     noMaxSpend: true,
-    noChangeMiningFee: true
+    noChangeMiningFee: true,
+    unstoppableDomainsTicker: 'HBAR'
   },
   polkadot: {
     initWalletName: lstrings.string_first_polkadot_wallet_name,
@@ -679,7 +713,8 @@ export const SPECIAL_CURRENCY_INFO: {
       modalMessage: lstrings.request_dot_minimum_notification_body,
       alertMessage: lstrings.request_dot_minimum_notification_alert_body
     },
-    isImportKeySupported: true
+    isImportKeySupported: true,
+    unstoppableDomainsTicker: 'DOT'
   },
   liberland: {
     initWalletName: lstrings.string_first_liberland_wallet_name,
@@ -693,7 +728,7 @@ export const SPECIAL_CURRENCY_INFO: {
       modalMessage: lstrings.request_lld_minimum_notification_body,
       alertMessage: lstrings.request_lld_minimum_notification_alert_body
     },
-    isTransactionListUnsupported: true,
+    isTransactionListUnsupported: false,
     isImportKeySupported: true
   },
   liberlandtestnet: {
@@ -735,7 +770,8 @@ export const SPECIAL_CURRENCY_INFO: {
         inputType: 'number-pad',
         inputValidation: (input: string) => /^\d+$/.test(input) && gte(input, '419200') // sapling activation height
       }
-    ]
+    ],
+    unstoppableDomainsTicker: 'ZEC'
   },
   piratechain: {
     initWalletName: lstrings.string_first_piratechain_wallet_name,
@@ -756,7 +792,8 @@ export const SPECIAL_CURRENCY_INFO: {
         inputType: 'number-pad',
         inputValidation: (input: string) => /^\d+$/.test(input) && gte(input, '152855') // sapling activation height
       }
-    ]
+    ],
+    unstoppableDomainsTicker: 'ARRR'
   },
   polygon: {
     initWalletName: lstrings.string_first_polygon_wallet_name,
@@ -768,7 +805,8 @@ export const SPECIAL_CURRENCY_INFO: {
     walletConnectV2ChainId: {
       namespace: 'eip155',
       reference: '137'
-    }
+    },
+    unstoppableDomainsTicker: 'MATIC'
   },
   pulsechain: {
     initWalletName: lstrings.string_first_pulsechain_wallet_name,
@@ -793,7 +831,8 @@ export const SPECIAL_CURRENCY_INFO: {
     walletConnectV2ChainId: {
       namespace: 'eip155',
       reference: '43114'
-    }
+    },
+    unstoppableDomainsTicker: 'AVAX'
   },
   algorand: {
     initWalletName: lstrings.string_first_algorand_wallet_name,
@@ -805,7 +844,8 @@ export const SPECIAL_CURRENCY_INFO: {
     walletConnectV2ChainId: {
       namespace: 'algorand',
       reference: 'wGHE2Pwdvd7S12BL5FaOP20EGYesN73k'
-    }
+    },
+    unstoppableDomainsTicker: 'ALGO'
   },
   holesky: {
     initWalletName: lstrings.string_first_holesky_wallet_name,