+++ src/plugins/stake-plugins/generic/policyAdapters/EthereumKilnAdaptor.ts
@@ -68,8 +68,13 @@ export const makeEthereumKilnAdapter = (policyConfig: StakePolicyConfig<Ethereum
     const walletSigner = new EdgeWalletSigner(wallet, provider)
     const walletAddress = await walletSigner.getAddress()
 
-    let txCount: number = await walletSigner.getTransactionCount('pending')
-    const nextNonce = (): number => txCount++
+    let txCount: number | undefined
+    const nextNonce = async (): Promise<number> => {
+      if (txCount == null) {
+        txCount = await walletSigner.getTransactionCount('pending')
+      }
+      return txCount++
+    }
 
     const feeData = await provider.getFeeData()
     const maxFeePerGas = feeData.maxFeePerGas !== null ? feeData.maxFeePerGas : undefined
@@ -113,7 +118,7 @@ export const makeEthereumKilnAdapter = (policyConfig: StakePolicyConfig<Ethereum
         gasLimit: '500000',
         maxFeePerGas,
         maxPriorityFeePerGas,
-        nonce: nextNonce(),
+        nonce: await nextNonce(),
         customData: {
           metadata: {
             name: metadataName,
@@ -133,7 +138,7 @@ export const makeEthereumKilnAdapter = (policyConfig: StakePolicyConfig<Ethereum
         gasLimit: '250000', // Typically uses 190000-225000 gas
         maxFeePerGas,
         maxPriorityFeePerGas,
-        nonce: nextNonce(),
+        nonce: await nextNonce(),
         value: requestNativeAmount,
         customData: {
           metadata: {
@@ -163,7 +168,7 @@ export const makeEthereumKilnAdapter = (policyConfig: StakePolicyConfig<Ethereum
         gasLimit: '500000',
         maxFeePerGas,
         maxPriorityFeePerGas,
-        nonce: nextNonce(),
+        nonce: await nextNonce(),
         customData: {
           metadata: {
             name: metadataName,